refactor: Special List View Model
