+++ frontend-hybrid/yakal/lib/models/Profile/special_note_model.dart
